pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "computer-use-mcp"
version = "0.1.7"
description = "MCP server that provides computer control capabilities using PyAutoGUI, RapidOCR, ONNXRuntime. Similar to 'computer-use' by Anthropic. With Zero External Dependencies."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities"
]
dependencies = [
    "python-Levenshtein",
    "pyautogui",
    "mcp[cli]",
    "pillow",
    "pygetwindow",
    "fuzzywuzzy",
    "rapidocr",
    "onnxruntime",
    "rapidocr_onnxruntime",
    "opencv-python"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[project.urls]
Homepage = "https://github.com/zhouyuf6741/computer-use-mcp.git"
Documentation = "https://github.com/zhouyuf6741/computer-use-mcp#readme"

[project.scripts]
computer-use-mcp = "computer_use_mcp.cli:main"
computer-use-mcp-server = "computer_use_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/computer_use_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.hatch.build]
sources = ["src"]
packages = ["src/computer_use_mcp"]

